Convert Cobbler Profile virtFileSize from int to double

+++ java/code/src/com/redhat/rhn/frontend/dto/kickstart/KickstartDto.java
@@ -36,7 +36,7 @@ public class KickstartDto extends BaseDto {
     private String cobblerId;
     private String cobblerUrl;
     private int virtMemory;
-    private int virtSpace;
+    private double virtSpace;
     private int virtCpus;
     private String virtBridge;
     private String macAddress;
@@ -217,7 +217,7 @@ public class KickstartDto extends BaseDto {
     /**
      * @return Returns the virtSpace.
      */
-    public int getVirtSpace() {
+    public double getVirtSpace() {
         return virtSpace;
     }
 
@@ -225,7 +225,7 @@ public class KickstartDto extends BaseDto {
     /**
      * @param virtSpaceIn The virtSpace to set.
      */
-    public void setVirtSpace(int virtSpaceIn) {
+    public void setVirtSpace(double virtSpaceIn) {
         this.virtSpace = virtSpaceIn;
     }
